+++ src/rwkv_tl/state.py
@@ -32,7 +32,7 @@ class State(tuple[list[dict[str, Tensor]], list[dict[str, Tensor]]]):
         for _ in range(n_layer):
             tmix_layers.append(
                 {
-                    "x": torch.zeros(n_embd, dtype=torch.bfloat16, device=device),
+                    "x": torch.zeros(n_embd, dtype=torch.float16, device=device),
                     "rnn": torch.zeros(
                         (head_count, head_dim, head_dim),
                         dtype=torch.float32,
@@ -41,7 +41,7 @@ class State(tuple[list[dict[str, Tensor]], list[dict[str, Tensor]]]):
                 }
             )
             cmix_layers.append(
-                {"x": torch.zeros(n_embd, dtype=torch.bfloat16, device=device)}
+                {"x": torch.zeros(n_embd, dtype=torch.float16, device=device)}
             )
 
         obj = super().__new__(cls, (tmix_layers, cmix_layers))
